requirements.txt
requests~=2.28.2
python-decouple~=3.8
pandas~=2.0.0